fix a bunch of warnings (could actually be bugs)

+++ src/iso9660.h
@@ -165,9 +165,9 @@ public:
 
         ISO_FSEEK(fp, (long long)lba * 2048, SEEK_SET);
         uint8_t buf[65536];
-        uint32_t remaining = size;
+        SSIZE_T remaining = size;
         while (remaining > 0) {
-            uint32_t chunk = remaining < sizeof(buf) ? remaining : sizeof(buf);
+            SSIZE_T chunk = remaining < sizeof(buf) ? remaining : sizeof(buf);
             size_t rd = fread(buf, 1, chunk, fp);
             if (rd == 0) break;
             fwrite(buf, 1, rd, out);
